Add login authentication and api
