+++ .github/workflows/linters.yml
@@ -20,7 +20,7 @@ jobs:
           [ -f .eslintrc.json ] || wget https://raw.githubusercontent.com/microverseinc/linters-config/master/react-redux/.eslintrc.json
           [ -f .babelrc ] || wget https://raw.githubusercontent.com/microverseinc/linters-config/master/react-redux/.babelrc
       - name: ESLint Report
-        run: npx eslint .
+        run: npx eslint . --fix
   stylelint:
     name: Stylelint
     runs-on: ubuntu-18.04
@@ -34,7 +34,7 @@ jobs:
           npm install --save-dev stylelint@13.x stylelint-scss@3.x stylelint-config-standard@21.x stylelint-csstree-validator@1.x
           [ -f .stylelintrc.json ] || wget https://raw.githubusercontent.com/microverseinc/linters-config/master/react-redux/.stylelintrc.json
       - name: Stylelint Report
-        run: npx stylelint "**/*.{css,scss}"
+        run: npx stylelint "**/*.{css,scss}" --fix
   nodechecker:
     name: node_modules checker
     runs-on: ubuntu-18.04
