+++ package.json
@@ -7,6 +7,7 @@
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
+        "axios": "^1.1.3",
         "espree": "^9.4.0",
         "firebase": "^9.14.0",
         "prop-types": "^15.8.1",
@@ -48,10 +49,10 @@
         ]
     },
     "devDependencies": {
-        "@babel/core": "^7.19.6",
+        "@babel/core": "^7.20.2",
         "@babel/eslint-parser": "^7.19.1",
         "@babel/plugin-syntax-jsx": "^7.18.6",
-        "@babel/preset-env": "^7.19.4",
+        "@babel/preset-env": "^7.20.2",
         "@babel/preset-react": "^7.18.6",
         "eslint": "^7.32.0",
         "eslint-config-airbnb": "^18.2.1",
