+++ src/redux/auth/Api.js
@@ -1,0 +1,26 @@
+// import axios from 'axios';
+
+export const URL_LOGIN = 'https://protected-sea-38971.herokuapp.com/auth/login';
+export const URL_SIGNUP = 'https://protected-sea-38971.herokuapp.com/api/v1/users';
+
+// export const login = async (params) => axios({
+//   method: 'post',
+//   url: `${URL_LOGIN}/sessions`,
+//   data: {
+//     email: params.email,
+//     password: params.password,
+//   },
+//   withCredentials: true,
+// });
+
+// export const signUp = async (params) => axios({
+//   method: 'post',
+//   url: `${URL_SIGNUP}/registration`,
+//   data: {
+//     name: params.name,
+//     email: params.email,
+//     password: params.password,
+//     password_confirmation: params.password_confirmation,
+//   },
+//   withCredentials: true,
+// });
